+++ src/utils/compose.js
@@ -6,7 +6,7 @@
 import noop from 'utils/noop'
 
 /**
- * @see modules/README.md for the use of the mixin function
+ * @see modules/README.md for the use of the compose function
  * In essence, the first argument can be a function that simply returns a class
  * (this allows that class to extend whatever it likes),
  * although all other argument must return a function that extends the its
@@ -16,7 +16,7 @@ import noop from 'utils/noop'
  * nothing or whatever it likes whereas subsequent classes must inherit the
  * previous class in the list.
  */
-export default function mixin( ...modules ) {
+export default function compose( ...modules ) {
     // Add methods here to avoid throwing errors
     var base = class {
         update() {}