+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#
+
+*
+
 # 0.0.4
 
 * _update_ refactor entities and entity inheritance chain